refactor : oauthInfo 디렉토리명 수정
